+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All feature additions, significant bug fixes, and API changes will be documented
 in this file. This project follows [semantic versioning](https://semver.org/).
 
+# 13.1
+- Add the `distinctUntilChanged` operator to `Observable`.
+- Add the `distinctUntilChangedBy` operator to `Observable`.
+
 # 13.0
 - Update the `Pool` to solve various bugs, to allocate into block sizes based on
   best fit for ag iven type, and to allocate memory in large chunks for better